:recycle: refactor: Passando para unica classe
