--- src/main/java/br/com/edu/ifmt/sacode/domain/services/GetAllUserService.java
@@ -1,4 +1,0 @@
-package br.com.edu.ifmt.sacode.domain.services;
-
-public class GetAllUserService {
-}
